+++ src/main/java/org/zalando/nakadi/controller/ExceptionHandling.java
@@ -10,21 +10,30 @@ import org.springframework.http.converter.HttpMessageNotReadableException;
 import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.context.request.NativeWebRequest;
-import org.zalando.nakadi.exceptions.runtime.FeatureNotAvailableException;
-import org.zalando.nakadi.exceptions.runtime.IllegalClientIdException;
-import org.zalando.nakadi.exceptions.NakadiException;
-import org.zalando.nakadi.exceptions.NakadiRuntimeException;
-import org.zalando.nakadi.exceptions.runtime.TimelineException;
 import org.zalando.nakadi.exceptions.runtime.AccessDeniedException;
+import org.zalando.nakadi.exceptions.runtime.CompactionException;
 import org.zalando.nakadi.exceptions.runtime.CursorConversionException;
 import org.zalando.nakadi.exceptions.runtime.CursorsAreEmptyException;
 import org.zalando.nakadi.exceptions.runtime.DbWriteOperationsBlockedException;
+import org.zalando.nakadi.exceptions.runtime.EnrichmentException;
+import org.zalando.nakadi.exceptions.runtime.FeatureNotAvailableException;
+import org.zalando.nakadi.exceptions.runtime.IllegalClientIdException;
+import org.zalando.nakadi.exceptions.runtime.InternalNakadiException;
+import org.zalando.nakadi.exceptions.runtime.InvalidPartitionKeyFieldsException;
 import org.zalando.nakadi.exceptions.runtime.LimitReachedException;
-import org.zalando.nakadi.exceptions.runtime.MyNakadiRuntimeException1;
-import org.zalando.nakadi.exceptions.runtime.NoEventTypeException;
+import org.zalando.nakadi.exceptions.runtime.NakadiBaseException;
+import org.zalando.nakadi.exceptions.runtime.NakadiRuntimeException;
+import org.zalando.nakadi.exceptions.runtime.NoStreamingSlotsAvailable;
+import org.zalando.nakadi.exceptions.runtime.NoSuchEventTypeException;
+import org.zalando.nakadi.exceptions.runtime.NoSuchPartitionStrategyException;
+import org.zalando.nakadi.exceptions.runtime.NoSuchSchemaException;
+import org.zalando.nakadi.exceptions.runtime.NoSuchSubscriptionException;
+import org.zalando.nakadi.exceptions.runtime.PartitioningException;
 import org.zalando.nakadi.exceptions.runtime.RepositoryProblemException;
 import org.zalando.nakadi.exceptions.runtime.ServiceTemporarilyUnavailableException;
+import org.zalando.nakadi.exceptions.runtime.TimelineException;
 import org.zalando.nakadi.exceptions.runtime.TopicCreationException;
+import org.zalando.nakadi.exceptions.runtime.UnprocessableSubscriptionException;
 import org.zalando.problem.MoreStatus;
 import org.zalando.problem.Problem;
 import org.zalando.problem.spring.web.advice.ProblemHandling;
@@ -32,6 +41,9 @@ import org.zalando.problem.spring.web.advice.Responses;
 
 import javax.ws.rs.core.Response;
 
+import static javax.ws.rs.core.Response.Status.CONFLICT;
+import static javax.ws.rs.core.Response.Status.INTERNAL_SERVER_ERROR;
+import static javax.ws.rs.core.Response.Status.NOT_FOUND;
 import static javax.ws.rs.core.Response.Status.NOT_IMPLEMENTED;
 import static org.zalando.problem.MoreStatus.UNPROCESSABLE_ENTITY;
 
@@ -77,12 +89,6 @@ public final class ExceptionHandling implements ProblemHandling {
         return Responses.create(Response.Status.BAD_REQUEST, message, request);
     }
 
-    @ExceptionHandler(NoEventTypeException.class)
-    public ResponseEntity<Problem> noEventTypeException(final NoEventTypeException exception,
-                                                        final NativeWebRequest request) {
-        return Responses.create(Response.Status.NOT_FOUND, exception.getMessage(), request);
-    }
-
     @ExceptionHandler(AccessDeniedException.class)
     public ResponseEntity<Problem> accessDeniedException(final AccessDeniedException exception,
                                                          final NativeWebRequest request) {
@@ -106,9 +112,8 @@ public final class ExceptionHandling implements ProblemHandling {
     public ResponseEntity<Problem> handleExceptionWrapper(final NakadiRuntimeException exception,
                                                           final NativeWebRequest request) throws Exception {
         final Throwable cause = exception.getCause();
-        if (cause instanceof NakadiException) {
-            final NakadiException ne = (NakadiException) cause;
-            return Responses.create(ne.asProblem(), request);
+        if (cause instanceof InternalNakadiException) {
+            return Responses.create(INTERNAL_SERVER_ERROR, exception.getMessage(), request);
         }
         throw exception.getException();
     }
@@ -120,8 +125,8 @@ public final class ExceptionHandling implements ProblemHandling {
         return Responses.create(Response.Status.SERVICE_UNAVAILABLE, exception.getMessage(), request);
     }
 
-    @ExceptionHandler(MyNakadiRuntimeException1.class)
-    public ResponseEntity<Problem> handleInternalError(final MyNakadiRuntimeException1 exception,
+    @ExceptionHandler(NakadiBaseException.class)
+    public ResponseEntity<Problem> handleInternalError(final NakadiBaseException exception,
                                                        final NativeWebRequest request) {
         LOG.error("Unexpected problem occurred", exception);
         return Responses.create(Response.Status.INTERNAL_SERVER_ERROR, exception.getMessage(), request);
@@ -132,9 +137,8 @@ public final class ExceptionHandling implements ProblemHandling {
                                                            final NativeWebRequest request) {
         LOG.error(exception.getMessage(), exception);
         final Throwable cause = exception.getCause();
-        if (cause instanceof NakadiException) {
-            final NakadiException ne = (NakadiException) cause;
-            return Responses.create(ne.asProblem(), request);
+        if (cause instanceof InternalNakadiException) {
+            return Responses.create(Problem.valueOf(INTERNAL_SERVER_ERROR, exception.getMessage()), request);
         }
         return Responses.create(Response.Status.SERVICE_UNAVAILABLE, exception.getMessage(), request);
     }
@@ -179,9 +183,80 @@ public final class ExceptionHandling implements ProblemHandling {
     public ResponseEntity<Problem> handleFeatureNotAvailable(
             final FeatureNotAvailableException ex,
             final NativeWebRequest request) {
-        LOG.debug(ex.getMessage(), ex);
+        LOG.debug(ex.getMessage());
         return Responses.create(Problem.valueOf(NOT_IMPLEMENTED, ex.getMessage()), request);
-
     }
 
+    @ExceptionHandler(CompactionException.class)
+    public ResponseEntity<Problem> handleCompactionException(final CompactionException exception,
+                                                             final NativeWebRequest request) {
+        LOG.debug(exception.getMessage());
+        return Responses.create(UNPROCESSABLE_ENTITY, exception.getMessage(), request);
+    }
+
+    @ExceptionHandler(EnrichmentException.class)
+    public ResponseEntity<Problem> handleEnrichmentException(final EnrichmentException exception,
+                                                             final NativeWebRequest request) {
+        LOG.debug(exception.getMessage());
+        return Responses.create(UNPROCESSABLE_ENTITY, exception.getMessage(), request);
+    }
+
+    @ExceptionHandler(NoSuchPartitionStrategyException.class)
+    public ResponseEntity<Problem> handleNoSuchPartitionStrategyException(
+            final NoSuchPartitionStrategyException exception,
+            final NativeWebRequest request) {
+        LOG.debug(exception.getMessage());
+        return Responses.create(UNPROCESSABLE_ENTITY, exception.getMessage(), request);
+    }
+
+    @ExceptionHandler(PartitioningException.class)
+    public ResponseEntity<Problem> handlePartitioningException(final PartitioningException exception,
+                                                               final NativeWebRequest request) {
+        LOG.debug(exception.getMessage());
+        return Responses.create(UNPROCESSABLE_ENTITY, exception.getMessage(), request);
+    }
+
+    @ExceptionHandler(InvalidPartitionKeyFieldsException.class)
+    public ResponseEntity<Problem> handleInvalidPartitionKeyFieldsException(
+            final InvalidPartitionKeyFieldsException exception,
+            final NativeWebRequest request) {
+        LOG.debug(exception.getMessage());
+        return Responses.create(UNPROCESSABLE_ENTITY, exception.getMessage(), request);
+    }
+
+    @ExceptionHandler(NoSuchEventTypeException.class)
+    public ResponseEntity<Problem> handleNoSuchEventTypeException(final NoSuchEventTypeException exception,
+                                                               final NativeWebRequest request) {
+        LOG.debug(exception.getMessage());
+        return Responses.create(NOT_FOUND, exception.getMessage(), request);
+    }
+
+    @ExceptionHandler(NoSuchSchemaException.class)
+    public ResponseEntity<Problem> handleNoSuchSchemaException(final NoSuchSchemaException exception,
+                                                               final NativeWebRequest request) {
+        LOG.debug(exception.getMessage());
+        return Responses.create(NOT_FOUND, exception.getMessage(), request);
+    }
+
+    @ExceptionHandler(NoSuchSubscriptionException.class)
+    public ResponseEntity<Problem> handleNoSuchSubscriptionException(final NoSuchSubscriptionException exception,
+                                                                     final NativeWebRequest request) {
+        LOG.debug(exception.getMessage());
+        return Responses.create(NOT_FOUND, exception.getMessage(), request);
+    }
+
+    @ExceptionHandler(NoStreamingSlotsAvailable.class)
+    public ResponseEntity<Problem> handleNoStreamingSlotsAvailable(final NoStreamingSlotsAvailable exception,
+                                                                   final NativeWebRequest request) {
+        LOG.debug(exception.getMessage());
+        return Responses.create(CONFLICT, exception.getMessage(), request);
+    }
+
+    @ExceptionHandler(UnprocessableSubscriptionException.class)
+    public ResponseEntity<Problem> handleUnprocessableSubscriptionException(
+            final UnprocessableSubscriptionException exception,
+            final NativeWebRequest request) {
+        LOG.debug(exception.getMessage());
+        return Responses.create(UNPROCESSABLE_ENTITY, exception.getMessage(), request);
+    }
 }