+++ src/main/java/org/zalando/nakadi/exceptions/CompactionException.java
@@ -1,9 +1,0 @@
-package org.zalando.nakadi.exceptions;
-
-public class CompactionException extends UnprocessableEntityException {
-
-    public CompactionException(final String message) {
-        super(message);
-    }
-
-}