+++ src/main/java/org/zalando/nakadi/exceptions/Try.java
@@ -1,5 +1,8 @@
 package org.zalando.nakadi.exceptions;
 
+import org.zalando.nakadi.exceptions.runtime.NakadiBaseException;
+import org.zalando.nakadi.exceptions.runtime.NakadiRuntimeException;
+
 import java.util.Optional;
 import java.util.concurrent.Callable;
 import java.util.function.Function;
@@ -95,6 +98,8 @@ public abstract class Try<T> {
         public T getOrThrow() {
             if (exception instanceof NakadiRuntimeException) {
                 throw (NakadiRuntimeException) exception;
+            } else if (exception instanceof NakadiBaseException) {
+                throw (NakadiBaseException) exception;
             }
             throw new NakadiRuntimeException(exception);
         }