+++ src/main/java/org/zalando/nakadi/exceptions/runtime/TimeLagStatsTimeoutException.java
@@ -1,6 +1,6 @@
 package org.zalando.nakadi.exceptions.runtime;
 
-public class TimeLagStatsTimeoutException extends MyNakadiRuntimeException1 {
+public class TimeLagStatsTimeoutException extends NakadiBaseException {
 
     public TimeLagStatsTimeoutException(final String msg, final Throwable cause) {
         super(msg, cause);