+++ src/main/java/org/zalando/nakadi/exceptions/runtime/TopicConfigException.java
@@ -1,6 +1,6 @@
 package org.zalando.nakadi.exceptions.runtime;
 
-public class TopicConfigException extends MyNakadiRuntimeException1 {
+public class TopicConfigException extends NakadiBaseException {
 
     public TopicConfigException(final String message, final Exception e) {
         super(message, e);