+++ src/main/java/org/zalando/nakadi/repository/EventConsumer.java
@@ -4,8 +4,7 @@ import org.zalando.nakadi.domain.ConsumedEvent;
 import org.zalando.nakadi.domain.EventTypePartition;
 import org.zalando.nakadi.domain.NakadiCursor;
 import org.zalando.nakadi.domain.TopicPartition;
-import org.zalando.nakadi.exceptions.InvalidCursorException;
-import org.zalando.nakadi.exceptions.NakadiException;
+import org.zalando.nakadi.exceptions.runtime.InvalidCursorException;
 
 import java.io.Closeable;
 import java.util.Collection;
@@ -23,6 +22,6 @@ public interface EventConsumer extends Closeable {
     interface ReassignableEventConsumer extends EventConsumer {
         Set<EventTypePartition> getAssignment();
 
-        void reassign(Collection<NakadiCursor> newValues) throws NakadiException, InvalidCursorException;
+        void reassign(Collection<NakadiCursor> newValues) throws InvalidCursorException;
     }
 }