+++ src/main/java/org/zalando/nakadi/repository/db/SubscriptionDbRepository.java
@@ -18,10 +18,9 @@ import org.springframework.jdbc.core.RowMapper;
 import org.springframework.stereotype.Component;
 import org.zalando.nakadi.domain.Subscription;
 import org.zalando.nakadi.domain.SubscriptionBase;
-import org.zalando.nakadi.exceptions.NoSuchSubscriptionException;
 import org.zalando.nakadi.exceptions.runtime.DuplicatedSubscriptionException;
 import org.zalando.nakadi.exceptions.runtime.InconsistentStateException;
-import org.zalando.nakadi.exceptions.runtime.NoSubscriptionException;
+import org.zalando.nakadi.exceptions.runtime.NoSuchSubscriptionException;
 import org.zalando.nakadi.exceptions.runtime.RepositoryProblemException;
 import org.zalando.nakadi.exceptions.runtime.ServiceTemporarilyUnavailableException;
 import org.zalando.nakadi.util.HashGenerator;
@@ -161,7 +160,7 @@ public class SubscriptionDbRepository extends AbstractDbRepository {
 
     public Subscription getSubscription(final String owningApplication, final Set<String> eventTypes,
                                         final String consumerGroup)
-            throws InconsistentStateException, NoSubscriptionException, RepositoryProblemException {
+            throws InconsistentStateException, NoSuchSubscriptionException, RepositoryProblemException {
 
         final String sql = "SELECT s_subscription_object FROM zn_data.subscription " +
                 "WHERE s_subscription_object->>'owning_application' = ? " +
@@ -174,7 +173,7 @@ public class SubscriptionDbRepository extends AbstractDbRepository {
         } catch (final JsonProcessingException e) {
             throw new InconsistentStateException("Deserialization problem during reading subscription from DB", e);
         } catch (final EmptyResultDataAccessException e) {
-            throw new NoSubscriptionException("Subscription does not exist", e);
+            throw new NoSuchSubscriptionException("Subscription does not exist", e);
         } catch (final DataAccessException e) {
             throw new RepositoryProblemException("Error occurred when reading subscription from DB", e);
         }