+++ src/main/java/org/zalando/nakadi/service/EventTypeService.java
@@ -25,12 +25,6 @@ import org.zalando.nakadi.domain.EventTypeStatistics;
 import org.zalando.nakadi.domain.Subscription;
 import org.zalando.nakadi.domain.Timeline;
 import org.zalando.nakadi.enrichment.Enrichment;
-import org.zalando.nakadi.exceptions.InternalNakadiException;
-import org.zalando.nakadi.exceptions.NakadiException;
-import org.zalando.nakadi.exceptions.NakadiRuntimeException;
-import org.zalando.nakadi.exceptions.NoSuchEventTypeException;
-import org.zalando.nakadi.exceptions.NoSuchPartitionStrategyException;
-import org.zalando.nakadi.exceptions.NoSuchSubscriptionException;
 import org.zalando.nakadi.exceptions.runtime.AccessDeniedException;
 import org.zalando.nakadi.exceptions.runtime.ConflictException;
 import org.zalando.nakadi.exceptions.runtime.DbWriteOperationsBlockedException;
@@ -40,8 +34,12 @@ import org.zalando.nakadi.exceptions.runtime.EventTypeOptionsValidationException
 import org.zalando.nakadi.exceptions.runtime.EventTypeUnavailableException;
 import org.zalando.nakadi.exceptions.runtime.FeatureNotAvailableException;
 import org.zalando.nakadi.exceptions.runtime.InconsistentStateException;
+import org.zalando.nakadi.exceptions.runtime.InternalNakadiException;
 import org.zalando.nakadi.exceptions.runtime.InvalidEventTypeException;
-import org.zalando.nakadi.exceptions.runtime.NoEventTypeException;
+import org.zalando.nakadi.exceptions.runtime.NakadiRuntimeException;
+import org.zalando.nakadi.exceptions.runtime.NoSuchEventTypeException;
+import org.zalando.nakadi.exceptions.runtime.NoSuchPartitionStrategyException;
+import org.zalando.nakadi.exceptions.runtime.NoSuchSubscriptionException;
 import org.zalando.nakadi.exceptions.runtime.NotFoundException;
 import org.zalando.nakadi.exceptions.runtime.ServiceTemporarilyUnavailableException;
 import org.zalando.nakadi.exceptions.runtime.TimelineException;
@@ -61,6 +59,7 @@ import org.zalando.nakadi.service.validation.EventTypeOptionsValidator;
 import org.zalando.nakadi.util.JsonUtils;
 import org.zalando.nakadi.validation.SchemaEvolutionService;
 import org.zalando.nakadi.validation.SchemaIncompatibility;
+import org.zalando.problem.Problem;
 
 import java.io.Closeable;
 import java.io.IOException;
@@ -71,6 +70,8 @@ import java.util.Optional;
 import java.util.concurrent.TimeoutException;
 import java.util.stream.Collectors;
 
+import static javax.ws.rs.core.Response.Status.INTERNAL_SERVER_ERROR;
+import static javax.ws.rs.core.Response.Status.NOT_FOUND;
 import static org.zalando.nakadi.service.FeatureToggleService.Feature.DELETE_EVENT_TYPE_WITH_SUBSCRIPTIONS;
 
 @Component
@@ -215,7 +216,7 @@ public class EventTypeService {
     }
 
     public void delete(final String eventTypeName) throws EventTypeDeletionException, AccessDeniedException,
-            NoEventTypeException, ConflictException, ServiceTemporarilyUnavailableException,
+            NoSuchEventTypeException, ConflictException, ServiceTemporarilyUnavailableException,
             DbWriteOperationsBlockedException {
         if (featureToggleService.isFeatureEnabled(FeatureToggleService.Feature.DISABLE_DB_WRITE_OPERATIONS)) {
             throw new DbWriteOperationsBlockedException("Cannot delete event type: write operations on DB " +
@@ -229,7 +230,7 @@ public class EventTypeService {
 
             final Optional<EventType> eventTypeOpt = eventTypeRepository.findByNameO(eventTypeName);
             if (!eventTypeOpt.isPresent()) {
-                throw new NoEventTypeException("EventType \"" + eventTypeName + "\" does not exist.");
+                throw new NoSuchEventTypeException("EventType \"" + eventTypeName + "\" does not exist.");
             }
             eventType = eventTypeOpt.get();
 
@@ -249,7 +250,7 @@ public class EventTypeService {
             LOG.error("Failed to wait for timeline switch", e);
             throw new EventTypeUnavailableException("Event type " + eventTypeName
                     + " is currently in maintenance, please repeat request");
-        } catch (final NakadiException | ServiceTemporarilyUnavailableException e) {
+        } catch (final InternalNakadiException | ServiceTemporarilyUnavailableException e) {
             LOG.error("Error deleting event type " + eventTypeName, e);
             throw new EventTypeDeletionException("Failed to delete event type " + eventTypeName);
         } finally {
@@ -355,7 +356,7 @@ public class EventTypeService {
             LOG.error("Failed to wait for timeline switch", e);
             throw new ServiceTemporarilyUnavailableException(
                     "Event type is currently in maintenance, please repeat request", e);
-        } catch (final NakadiException e) {
+        } catch (final InternalNakadiException e) {
             LOG.error("Unable to update event type", e);
             throw new NakadiRuntimeException(e);
         } finally {
@@ -375,7 +376,7 @@ public class EventTypeService {
                 .put("compatibility_mode", eventTypeBase.getCompatibilityMode()));
     }
 
-    private void updateRetentionTime(final EventType original, final EventType eventType) throws NakadiException {
+    private void updateRetentionTime(final EventType original, final EventType eventType) {
         final Long newRetentionTime = eventType.getOptions().getRetentionTime();
         final Long oldRetentionTime = original.getOptions().getRetentionTime();
         if (oldRetentionTime == null) {
@@ -399,13 +400,13 @@ public class EventTypeService {
     }
 
     private void updateEventTypeInDB(final EventType eventType, final Long newRetentionTime,
-                                     final Long oldRetentionTime) throws NakadiException {
-        final NakadiException exception = transactionTemplate.execute(action -> {
+                                     final Long oldRetentionTime) throws InternalNakadiException {
+        final InternalNakadiException exception = transactionTemplate.execute(action -> {
             try {
                 updateTimelinesCleanup(eventType.getName(), newRetentionTime, oldRetentionTime);
                 eventTypeRepository.update(eventType);
                 return null;
-            } catch (final NakadiException e) {
+            } catch (final InternalNakadiException e) {
                 return e;
             }
         });
@@ -444,10 +445,10 @@ public class EventTypeService {
             return Result.ok(eventType);
         } catch (final NoSuchEventTypeException e) {
             LOG.debug("Could not find EventType: {}", eventTypeName);
-            return Result.problem(e.asProblem());
+            return Result.problem(Problem.valueOf(NOT_FOUND, e.getMessage()));
         } catch (final InternalNakadiException e) {
             LOG.error("Problem loading event type " + eventTypeName, e);
-            return Result.problem(e.asProblem());
+            return Result.problem(Problem.valueOf(INTERNAL_SERVER_ERROR, e.getMessage()));
         }
     }
 
@@ -461,7 +462,7 @@ public class EventTypeService {
         } catch (TimelineException | NotFoundException e) {
             LOG.error("Problem deleting timeline for event type " + eventTypeName, e);
             throw new EventTypeDeletionException("Failed to delete timelines for event type " + eventTypeName);
-        } catch (NakadiException e) {
+        } catch (InternalNakadiException e) {
             LOG.error("Error deleting event type " + eventTypeName, e);
             throw new EventTypeDeletionException("Failed to delete event type " + eventTypeName);
         }